Move loadApplicationPermissions to DAO
